+++ src/builtin/exit.c
@@ -1,0 +1,3 @@
+#include "bshell.h"
+
+int bshell_exit(char **args) { return 0; }
