+++ src/builtin/reconfigure.c
@@ -1,0 +1,6 @@
+#include "bshell.h"
+
+int bshell_reconfigure() {
+  bshell_initialize();
+  return 1;
+}
